Revert "FEAT [usecroll, scrollSlice] muevo el hook que detecta si subes o baj…"
